# Conflicts:
#	login/src/main/java/io/bacta/login/server/rest/controller/GalaxiesController.java
#	login/src/main/java/io/bacta/login/server/service/DefaultGalaxyService.java
#	login/src/main/java/io/bacta/login/server/service/GalaxyNotFoundException.java
#	login/src/main/java/io/bacta/login/server/service/GalaxyService.java
#	postman/Bacta Rest API.postman_collection.json

+++ login/src/main/java/io/bacta/login/server/service/DefaultGalaxyService.java
@@ -125,8 +125,12 @@ public final class DefaultGalaxyService implements GalaxyService {
     }
 
     @Override
-    public void unregisterGalaxy(String name) {
+    public void unregisterGalaxy(String name) throws GalaxyNotFoundException {
         final Galaxy galaxy = galaxyRepository.findByName(name);
+
+        if (galaxy == null)
+            throw new GalaxyNotFoundException(name);
+
         galaxyRepository.delete(galaxy);
 
         //Unload it from memory if it exists there.
